select2 incorporado

+++ resources/views/sintomas/index.blade.php
@@ -198,40 +198,41 @@
 </script>
 
 <script type="text/javascript">
-function eliminar(id){
+function eliminar(id,sintoma){
+    if (confirm('¿Esta seguro de eliminar el sintoma '+sintoma+'? Tenga en cuenta que se eliminara todos los datos relacionados a ella.')) {
+          $.ajaxSetup({
+                        headers: {
+                            'X-CSRF-TOKEN': $('meta[name="csrf-token"]').attr('content')
+                        }
+            });
+          $.ajax({
 
-  $.ajaxSetup({
-                headers: {
-                    'X-CSRF-TOKEN': $('meta[name="csrf-token"]').attr('content')
-                }
-    });
-  $.ajax({
+                    type:'DELETE',
+                    url:"/sintomas/"+id,
+                    dataType:"json",
+                    data:{
+                        id:id,
+                    },
+                    success: function(response){
+                        var table=$("#myTable").DataTable();
+                        table.draw();
 
-            type:'DELETE',
-            url:"/sintomas/"+id,
-            dataType:"json",
-            data:{
-                id:id,
-            },
-            success: function(response){
-                var table=$("#myTable").DataTable();
-                table.draw();
+                        },
+                    error:function(err){
+                //        if (err.status == 422) { // when status code is 422, it's a validation issue
+                //     console.log(err.responseJSON);
+                //     $('#success_message').fadeIn().html(err.responseJSON.message);
 
-                },
-            error:function(err){
-        //        if (err.status == 422) { // when status code is 422, it's a validation issue
-        //     console.log(err.responseJSON);
-        //     $('#success_message').fadeIn().html(err.responseJSON.message);
-
-            
-        //     $.each(err.responseJSON.errors, function (i, error) {
-                
-        //         alert(error[0])
-        //     });
-        // }
-               
-            }
-        });
+                    
+                //     $.each(err.responseJSON.errors, function (i, error) {
+                        
+                //         alert(error[0])
+                //     });
+                // }
+                       
+                    }
+                });
+    }
 }
 
 </script>
